alerts: run when session detached
Recent changes meant that alerts would skip updating when the session was
detached.  However, with hooks, alerts should still be triggered in this
case.  It's possible to use if-shell with #{session_attached}, etc., to
check if the session is attached or not, should this be wanted.

Note that we still keep the session check when attaching, as this will still
keep the current behaviour.